pyproject.toml
[tool.poetry]
name = "timekpr-next-remote"
version = "0.1.0"
description = ""
authors = ["mrjones https://github.com/mrjones-plip"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fabric = "^3.2.2"
paramiko = "^3.5.0"
flask = "^3.1.0"
gotify = "^0.6.0"
humanize = "^4.11.0"
retry = "^0.9.2"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
livereload = "^2.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
